pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "spex"
version = "0.1.0"
description = ""
readme = "README.md"
authors = [{name = "Miles Hoar-Hewitt", email = "meles784@gmail.com"}]
requires-python = ">=3.13"

dependencies = [
    "pandas>=2.3.3",
    "requests>=2.32.5",
    "python-dotenv>=1.1.1",
    "openpyxl>=3.1.5"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.2"
]
web = [
    "fastapi>=0.119.1",
    "uvicorn>=0.38.0"
]

[tool.setuptools]
package-dir = {""="src"}

[tool.setuptools.packages.find]
where = ["src"]